# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateUserDto:
    properties:
      email:
        type: string
      fullName:
        minLength: 3
        type: string
      password:
        minLength: 4
        type: string
      phone:
        minLength: 10
        type: string
      role:
        enum:
        - admin
        - user
        - guest
        type: string
      status:
        enum:
        - active
        - inactive
        - deleted
        type: string
    required:
    - email
    - fullName
    - password
    - phone
    - role
    - status
    type: object
  dto.UpdateUserDto:
    properties:
      email:
        type: string
      fullName:
        minLength: 3
        type: string
      password:
        minLength: 4
        type: string
      phone:
        minLength: 10
        type: string
      role:
        enum:
        - admin
        - user
        - guest
        type: string
      status:
        enum:
        - active
        - inactive
        - deleted
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message: {}
      status:
        type: string
      statusCode:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: aimembabazi15@gmail.com
    name: aime
  description: API documentation for the User Service built with Go, Gin, and GORM.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Gin User Service API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid pagination parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all users with pagination
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user with full name, email, password, phone, status,
        and role.
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Marks a user as deleted without removing it from the database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: User already deleted
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Soft delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Fetch a user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a single user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user fields by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (partial allowed)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update user
      tags:
      - Users
  /users/{id}/restore:
    patch:
      consumes:
      - application/json
      description: Restore a user whose status is set to deleted
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User restored successfully
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: User is not deleted
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Restore a deleted user
      tags:
      - Users
schemes:
- http
swagger: "2.0"